% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{load_spss}
\alias{load_spss}
\title{Read SPSS file and parse factors from labels.}
\usage{
load_spss(path, cleanLabel = NULL, rm_notAnswered = TRUE,
  parse_frame_labels = TRUE)
}
\arguments{
\item{path}{(character), path of .sav file that should be loaded.}

\item{cleanLabel}{(function), pass through a function to clean the character
string of the column label.}

\item{rm_notAnswered}{(logical), if \code{TRUE} the 'Not Anwered' label will
not be included in factor levels if there are no instances.}

\item{parse_frame_labels}{(logical), if \code{TRUE}, will attempt to parse
"label" attribute into separate "label", "topic", and "category"
attributes based on string patterns in SPSS Labels exported by
Verint survey data.}
}
\value{
\code{data.frame} with factors and column attributes \code{label},
        \code{topic}, and \code{category}.
}
\description{
Reads an SPSS file with \code{sjmisc} and then converts
             columns to factors.
}
\note{
Labels and topics may be truncated if they are too long. This happens
      when the .sav file is generated by Verint. SPSS has a maximum length
      for variable labels, so those labels will be truncated in the SPSS file
      itself.
}
\examples{
file_path <- "dataset.sav"
df <- load_spss(file_path)


}

